privex-core/requirements-base.txt
# FastAPI & Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Data Processing
numpy>=1.26.0
pandas>=2.1.0

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
httpx>=0.27.0
pystray>=0.19.5
Pillow>=10.0.0
